added controller package and renamed packages.
added functionality to datasimulator class, not finished
