+++ backend-konijn/src/main/java/nl/cfns/base/BackendKonijnApplication.java
@@ -9,8 +9,8 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 import lombok.extern.java.Log;
 
-@SpringBootApplication(scanBasePackages = "nl.cfns.base, nl.cfns.configs, "
-		+ "nl.cfns.entities, " + "nl.cfns.services") // add
+@SpringBootApplication(scanBasePackages = "nl.cfns.base, nl.cfns.config, "
+		+ "nl.cfns.entity, " + "nl.cfns.service") // add
 																													// other
 																													// classes
 																													// to
@@ -22,9 +22,9 @@ import lombok.extern.java.Log;
 																													// enabled
 																													// here
 @Profile("dev") // profile for configurations. TODO: create diffent profile for production environment
-//@ComponentScan(basePackages = "nl.cfns.base, nl.cfns.configs, nl.cfns.entities") // add other classes to main app. autoconfiguration is automatically enabled here
-@EntityScan("nl.cfns.entities, nl.cfns.basicpojo") // add database object structures to main app
-@EnableJpaRepositories("nl.cfns.repositories")// find repository classes in repository package
+//@ComponentScan(basePackages = "nl.cfns.base, nl.cfns.config, nl.cfns.entity") // add other classes to main app. autoconfiguration is automatically enabled here
+@EntityScan("nl.cfns.entity, nl.cfns.basicpojo") // add database object structures to main app
+@EnableJpaRepositories("nl.cfns.repository")// find repository classes in repository package
 
 @Configuration //configuration annotation after scanning, this may prevent scanning problems
 @Log // to use logging in this class
