+++ backend-konijn/src/main/java/nl/cfns/config/package-info.java
@@ -1,0 +1,1 @@
+package nl.cfns.config;
