--- backend-konijn/src/main/java/nl/cfns/configs/package-info.java
@@ -1,1 +1,0 @@
-package nl.cfns.configs;
