+++ backend-konijn/src/main/java/nl/cfns/controller/HeatmapController.java
@@ -1,4 +1,4 @@
-package nl.cfns.base;
+package nl.cfns.controller;
 
 import java.io.IOException;
 import java.util.List;
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import nl.cfns.basicpojo.MeasurementPoint;
-import nl.cfns.services.HeatmapGenerator;
-import nl.cfns.services.SignalstrengthService;
+import nl.cfns.service.HeatmapGenerator;
+import nl.cfns.service.SignalstrengthService;
 
 @RestController
 @RequestMapping("/heatmap")
