+++ backend-konijn/src/main/java/nl/cfns/controller/package-info.java
@@ -1,0 +1,1 @@
+package nl.cfns.controller;
