+++ backend-konijn/src/main/java/nl/cfns/entity/Measurement.java
@@ -1,4 +1,4 @@
-package nl.cfns.entities;
+package nl.cfns.entity;
 
 import java.sql.Timestamp;
 
@@ -81,28 +81,7 @@ public class Measurement {
         this.id = null; // Set the current ID to null and generate a new one
     }
 	
-	//generate random values for testing purposes
-	public static Measurement generateRandomMeasurement() {
-	    Faker faker = new Faker();
-	    Measurement measurement = new Measurement();
-	    
-	    measurement.setId((long) 1);
-	    measurement.setTime(new Timestamp(System.currentTimeMillis())); 
-	    measurement.setLatency(faker.number().numberBetween(0, 100)); 
-	    measurement.setUpload((float) faker.number().randomDouble(2, 0, 100)); 
-	    measurement.setDownload((float) faker.number().randomDouble(2, 0, 100)); 
-	    measurement.setRSSI(faker.number().numberBetween(0, 100)); 
-	    measurement.setRSRQ(faker.number().numberBetween(0, 100)); 
-	    measurement.setRSRP(faker.number().numberBetween(0, 100)); 
-	    measurement.setSINR(faker.number().numberBetween(0, 100)); 
-	    measurement.setMnoString(faker.lorem().word()); 
-	    measurement.setLongitude(faker.number().randomDouble(2, 0, 100)); 
-	    measurement.setLatitude(faker.number().randomDouble(2, 0, 100)); 
-	    
-	    
-	    return measurement;
-	}
-	
+
 
 
 }
