+++ backend-konijn/src/main/java/nl/cfns/entity/Measuringbox2.java
@@ -1,4 +1,4 @@
-package nl.cfns.entities;
+package nl.cfns.entity;
 
 import com.github.javafaker.Faker;
 import jakarta.persistence.Column;
@@ -60,21 +60,6 @@ public class Measuringbox2 {
         this.id = null; // Set the current ID to null and generate a new one
     }
 	
-	
-	public static Measuringbox2 generateRandomMeasuringbox() {
-	    Faker faker = new Faker();
-	    Measuringbox2 measuringbox = new Measuringbox2();
-	    
-	    measuringbox.setId((long) 1);
-	    measuringbox.setMnc(faker.number().digits(3)); 
-	    measuringbox.setMcc(faker.number().digits(3)); 
-	    measuringbox.setLac(faker.number().digits(5)); 
-	    measuringbox.setLongitude(faker.number().randomDouble(2, 0, 100)); 
-	    measuringbox.setLatitude(faker.number().randomDouble(2, 0, 100)); 
-	    measuringbox.setStatus(Measuringbox2.MeasuringboxStatus.values()[faker.number().numberBetween(0, 4)]); 
-	    
-	    return measuringbox;
-	}
 	// private boolean measuringboxGender;
 
 	// other fields, getters and setters
