+++ backend-konijn/src/main/java/nl/cfns/repository/MeasuringboxRepository.java
@@ -1,11 +1,11 @@
-package nl.cfns.repositories;
+package nl.cfns.repository;
 
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import nl.cfns.entities.Measuringbox;
+import nl.cfns.entity.Measuringbox;
 
 //repository for measuringbox for interacting with the database
 @Repository("MeasuringboxRepository")
