+++ backend-konijn/src/main/java/nl/cfns/repository/package-info.java
@@ -1,3 +1,3 @@
-package nl.cfns.repositories;
+package nl.cfns.repository;
 
 //package for repositories to manipulate database. only contains simple save, delete and search by functions
