+++ backend-konijn/src/main/java/nl/cfns/service/package-info.java
@@ -1,3 +1,3 @@
-package nl.cfns.services;
+package nl.cfns.service;
 
 //contains logic for controllers. Implementation of the service layer. Uses repositories to manipulate database
